Initial function move
Moved a copy of handlebulkupload to the sponsor user management page. Would have liked this to be a more modular function (I would have implemented it as a helper method that took the user type's perms as a parameter) but well looks like Julia beat me to it so here we gooooo!