Removed more references to O uploads in the sponsorbulkloading variant.
Since this file is largely constructed from code copied from another file in our project, there are a lot of things that need to be excised, modified, or added before it can truly be called functioning as intended. It does save time overall, but it still takes a bit to write the file this way.